Fixes ```attempt to index field 'map' (a function value)```
If I'm not mistaken ```bindings.map``` refers to the function ```map``` itself, while ```bindings.map()``` refers to the result of the data. It's not possible to index the function itself, it is however possible to index the table which ```bindings.map``` returns i.e. ```bindings.map().insert(...)``` would work too.