Introducing a simple ranking algorithm
This commit introduces a new ranking algorithm: all the completion
candidates are first ranked by the weight of the client and only then
withing each specific client they are ranked according to a more
sophisticated metric.

This change allows one to have guaranteed consistent client ordering in the
completion results. This can be achieved simply by assigning appropriate
`weight_adjust` values to the clients. For example, setting
`weight_adjust = 2.0` next to the LSP client will guarantee that the LSP
results will be always on top.