Fix venv check
If the venv path is a symlink somewhere along the way, the comparison to the executable path can fail (e.g. if `XDG_DATA_HOME`/`vim.fn.stdpath("data")` is a symlink to another directory). Resolve the `_RT_PY` path the same way `_EXEC_PATH` is resolved so the comparison is valid